better types added to all packets

+++ netclient/c_lib/chat/packets.hpp
@@ -5,7 +5,7 @@
 class ChatMessage_CtoS: public FixedSizeReliableNetPacketToServer<ChatMessage_CtoS>
 {
     public:
-        int channel;
+        uint8_t channel;
         char msg[CHAT_MESSAGE_SIZE_MAX];
 
         inline void packet(char* buff, int* buff_n, bool pack)
@@ -19,8 +19,8 @@ class ChatMessage_CtoS: public FixedSizeReliableNetPacketToServer<ChatMessage_Ct
 class ChatMessage_StoC: public FixedSizeReliableNetPacketToClient<ChatMessage_StoC>
 {
     public:
-        int channel;
-        int sender;
+        uint8_t channel;
+        uint8_t sender;
         char msg[CHAT_MESSAGE_SIZE_MAX];
 
         inline void packet(char* buff, int* buff_n, bool pack)
